requirements.txt
# Python dependencies for Shipping a Data Product pipeline
# Pin versions as needed for reproducibility
python-dotenv~=1.0
telethon~=1.34
psycopg2-binary~=2.9
fastapi~=0.110
uvicorn[standard]~=0.29
dbt-core~=1.8
dbt-postgres~=1.8
dagster~=1.7
dagster-postgres~=0.23
ultralytics~=8.1
pillow~=10.3
requests~=2.32
pydantic~=2.7
pydantic-settings~=2.2
